almost exploded because I had a detached head and thought I lost the past 8 hours of work. Thank GOD FOR NVIM UNDO The detached git commit history had everything, but that's gone so let's write down what happened.
Created the pagination component and got it to properly link to other pages of the paginated book list
Actually styled the <Book> component, where now it is a card that has 3 tabs, one for general info (title, author, summary)
   one for publication info (ISBN, genres, langauges)
   one for API info, in this case the raw book object as retrieved from the server, as well as clickable API link to the API Detail Endpoint

I then significantly updated the BookList component to actually center the paginators, book cards, and title.
This had some difficulties, where I would put all my elements inside of a React-Boostrap <Stack> (basically a wrapper for a column flexbox)
but the actual cards would not be aligned.
I resorted to creating my own custom wrapper with styled components (literally called MyBookListWrapper) which had the desired forced width and flex-box alignment properties.

Combining all of these changes together, we now have a functioning, paginated, book List View where all data is gotten from the server!

There is no kind of local storage of information, which could be improved upon to make it faster.

Known visual bug is that the last page is very slightly shifted towards the right, which causes a jitter whenever you navigate to and from any page to the last page the page will "jitter"
AKA shift everything slightly over in a very obvious way.

My gut instinct is that this is something with how the pagination object is created, but I am not really sure.

Next step is the detail view of Books, which could then get linked up to the list view.